Refactor token handling for OAuth2 and M2M support
Refactored token generation and refresh logic to support OAuth2 and machine-to-machine (M2M) authentication. Removed deprecated methods and models related to legacy token handling. Introduced new request and response models, helper methods, and attributes for improved structure and maintainability. Updated headers, logging, and caching for better performance and traceability. Added stricter validation for attributes and enhanced expiration logic. Adjusted default configurations and constants to align with the new functionality.

+++ src/OpenSettings/Models/OpenSettingsConfigDataCacheModel.cs
@@ -6,7 +6,7 @@ namespace OpenSettings.Models
     {
         public T Data { get; set; }
 
-        public int ExpiresInSeconds { get; set; }
+        public double ExpiresInSeconds { get; set; }
 
         public DateTimeOffset AbsoluteExpiration { get; set; }
     }
