+++ src/OpenSettings/Models/OpenSettingsConfigModel.cs
@@ -4,7 +4,7 @@ namespace OpenSettings.Models
 {
     public class OpenSettingsConfigModel
     {
-        public int ExpiresInSeconds { get; set; }
+        public double ExpiresInSeconds { get; set; }
 
         public Dictionary<string, OpenSettingsConfigsDataModel> Data { get; set; } = new Dictionary<string, OpenSettingsConfigsDataModel>();
     }
